requirements.txt
awscli==1.6.2
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
bcdoc==0.12.2
blist==1.3.6
boto==2.33.0
botocore==0.73.0
cassandra-driver==3.12.0
certifi==2015.4.28
colorama==0.2.5
cql==1.4.0
cqlsh==5.0.4
decorator==4.0.9
docutils==0.14
enum34==1.1.6
functools32==3.2.3.post2
futures==3.1.1
genson==0.1.0
GraphLab-Create==2.1
GraphLab-Create-License==2.1
ipython==5.5.0
ipython-genutils==0.2.0
jmespath==0.5.0
jsonschema==2.5.1
multipledispatch==0.4.9
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
pkg-resources==0.0.0
prettytable==0.7.2
prompt-toolkit==1.0.15
psclient==2.0
ptyprocess==0.5.2
pyasn1==0.4.2
Pygments==2.2.0
python-dateutil==2.6.1
requests==2.9.1
rsa==3.1.2
scandir==1.6
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sseclient==0.0.8
thrift==0.10.0
tornado==4.3
traitlets==4.3.2
wcwidth==0.1.7
